+++ src/library.js
@@ -1,0 +1,9 @@
+export default function createElement(elementType, classNames = '', attributes = {}, innerHTML = '') {
+  const elementObject = document.createElement(elementType);
+  if (classNames) elementObject.classList.add(...(classNames.split(' ')));
+  Object.keys(attributes).forEach((attribute) => {
+    elementObject.setAttribute(attribute, attributes[attribute]);
+  });
+  elementObject.innerHTML = innerHTML;
+  return elementObject;
+}
